Flush after each keypress
Previously there was a bug in the event handling logic under the
following circumstances:

If a `gocui.View` that had `Editable` set to true and `Wrap` set
to false had content pasted into it from the system clipboard,
*and* that content was wider than the gocui.View, the pasted
content that extended beyond the border of the view would have a
space inserted between each character.

I was able to trace the problem to an accounting discrepancy in the
location of the view's cursor. Ordinary typed characters (not pasted)
had different values for their cursor X position than pasted characters.
This led me to examine how pasted content is handled differently than
typed content by the event logic, and I discovered that the gocui is
flushed after each typed character, but only after processing *all*
pasted characters.

I tried ensuring a flush after each pasted character as well, and the
problem disappeared. However, flushing is an expensive operation. I'm
confident that there's a less expensive way to trigger the accounting
logic that will handle the cursor position, and I will continue to
investigate that. In the meantime, this version of the code does
prevent the bug, even if it redraws the screen too often.